no message
Change-Id: I695210e0d12812e48faa312e71deb733e255ae3c (+1 squashed commit)
Squashed commits:
[59b951c] Added tests

Change-Id: I63da184f01e812c627c0669810d6b7daac8d71ee (+1 squashed commit)
Squashed commits:
[5f4a987] feat(ui): Using ThemeData in ContactDetails

Create a seperate ThemeData file for all contact widgets.
This theme will be "injected" at the gallery and should
eventually be part of the buildContext as the widgets
are converted to modules.

Change-Id: Ibf68c3cf7b4526b846550e40670812e8baa1a063 (+1 squashed commit)
Squashed commits:
[5f4a987] feat(ui): Using ThemeData in ContactDetails

Create a seperate ThemeData file for all contact widgets.
This theme will be "injected" at the gallery and should
eventually be part of the buildContext as the widgets
are converted to modules.

Change-Id: Ibf68c3cf7b4526b846550e40670812e8baa1a063